「✨」 feat(Raycasting): Now working :tada:

+++ src/raycasting/vectwo.c
@@ -6,7 +6,7 @@
 /*   By: adjoly <adjoly@student.42angouleme.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/26 15:47:29 by adjoly            #+#    #+#             */
-/*   Updated: 2024/11/02 14:25:24 by adjoly           ###   ########.fr       */
+/*   Updated: 2024/11/02 15:42:28 by adjoly           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,14 +24,14 @@ void	vec2_add(t_vec2 *first, t_vec2 second)
 	first->x += second.x;
 	first->y += second.y;
 }
-
+		// to be removed
 void	vec2_sub(t_vec2 *first, t_vec2 second)
 {
 	first->x -= second.x;
 	first->y -= second.y;
 }
 
-t_vec2	vec2_mul_ret(t_vec2 vec, uint16_t n)
+t_vec2	vec2_mul_ret(t_vec2 vec, float n)
 {
 	t_vec2	ret;
